# docs/ncc-api-swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "NCC - Client API",
    "description": "The NEM Community Client (NCC) provides an RESTful API to the wallet functionality. Thus enabling alternative front-end implementations be provided by the community.",
    "version": "0.6.37-BETA"
  },
  "host": "localhost:8989",
  "schemes": [
    "http"
  ],
  "basePath": "/ncc/api",
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/wallet/account/transaction/send": {
      "post": {
        "description": "Sends a new transaction (i.e., sending NEM, messages, assets).",
        "tags": [
          "Wallet"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "transaction",
            "in": "body",
            "description": "The transaction information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TransferSendRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The account and transactions information."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/signature/send": {
      "post": {
        "description": "ends a new multisig signature.",
        "tags": [
          "Wallet",
          "Account",
          "Multi-Signature"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The multisig signature transaction information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MultisigSignatureRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sending was successfull."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/modification/send": {
      "post": {
        "description": "Sends a new multisig modification",
        "tags": [
          "Wallet",
          "Account",
          "Multi-Signature"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "in": "body",
            "description": "The multisig account modification transaction information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/MultisigSignatureRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modification was successfull."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/transaction/validate": {
      "post": {
        "description": "Requests inspecting the transaction for validation purposes. The returned result will include:\n* The minimum fee for sending the transaction.\n* - A value indicating whether or not the recipient can receive encrypted messages.",
        "tags": [
          "Wallet",
          "Account"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "transaction",
            "in": "body",
            "description": "The transaction information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PartialTransferInformationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modification was successfull.",
            "schema": {
              "$ref": "#/definitions/PartialTransferInformationViewModel"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/signature/validate": {
      "post": {
        "description": "Request inspecting the multisig signature transaction for validation purposes. The returned result will include:\n* A minimum fee for creating the transaction.",
        "tags": [
          "Wallet",
          "Account",
          "Multi-Signature"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "transaction",
            "in": "body",
            "description": "The transaction information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PartialSignatureInformationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modification was successfull.",
            "schema": {
              "$ref": "#/definitions/PartialFeeInformationViewModel"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/modification/validate": {
      "post": {
        "description": "Request inspecting the multisig modification transaction for validation purposes. The returned result will include:\n* A minimum fee for creating modification transaction.",
        "tags": [
          "Wallet",
          "Account",
          "Multi-Signature"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "transaction",
            "in": "body",
            "description": "The transaction information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PartialSignatureInformationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Modification was successfull.",
            "schema": {
              "$ref": "#/definitions/PartialTransferInformationViewModel"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/remote/activate": {
      "post": {
        "description": "Announces activation of delegated harvesting account to the network (FEE: 1)",
        "tags": [
          "Wallet",
          "Account",
          "Harvest"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "transaction",
            "in": "body",
            "description": "The transaction information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TransferImportanceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Activation was successfull."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/remote/deactivate": {
      "post": {
        "description": "Announces deactivation of delegated harvesting account to the network (FEE: 1)",
        "tags": [
          "Wallet",
          "Account",
          "Harvest"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "transaction",
            "in": "body",
            "description": "The transaction information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TransferImportanceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deactivation was successfull."
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
    "/wallet/account/remote/validate": {
      "post": {
        "description": "Requests inspecting the importance transfer transaction for validation purposes. The returned result will include:\n* The minimum fee for sending the transaction.",
        "tags": [
          "Wallet",
          "Account",
          "Multi-Signature"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "transaction",
            "in": "body",
            "description": "The transaction information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PartialTransferInformationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The validation information.",
            "schema": {
              "$ref": "#/definitions/PartialTransferInformationViewModel"
            }
          },
          "400": {
            "description": "A bad request was sent.",
            "schema": {
              "$ref": "#/responses/BadRequest"
            }
          },
          "default": {
            "description": "Any other exception occured.",
            "schema": {
              "$ref": "#/responses/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponseHttp": {
      "properties": {
        "timeStamp": {
          "type": "integer",
          "description": "Point in time when error happened. Time is based on NEM epoch time."
        },
        "error": {
          "type": "string",
          "description": "Error message."
        },
        "message": {
          "type": "string",
          "description": "Additional message about the error."
        },
        "status": {
          "$ref": "#/definitions/HttpErrorCodes",
          "description": "Status of the error."
        }
      },
      "example": {
        "timeStamp": 4978518,
        "error": "Not Found",
        "message": null,
        "status": 404
      }
    },
    "ErrorResponseNcc": {
      "properties": {
        "timeStamp": {
          "type": "integer",
          "description": "Point in time when error happened. Time is based on NEM epoch time."
        },
        "error": {
          "type": "string",
          "description": "Error message."
        },
        "message": {
          "type": "string",
          "description": "Additional message about the error."
        },
        "status": {
          "$ref": "#/definitions/NccErrorCodes",
          "description": "Status of the error."
        }
      },
      "example": {
        "timeStamp": 4978518,
        "error": "Not Found",
        "message": null,
        "status": 404
      }
    },
    "NccErrorCodes": {
      "type": "integer",
      "format": "int32",
      "description": "Error codes that indicate the reason for failing:  \n* 106 - WALLET_IS_NOT_OPEN, The specified wallet is not open.\n* 126 - ADDRESS_BOOK_IS_NOT_OPEN, The specified address book is not open.\n* 202 - NO_PUBLIC_KEY, The recipient public key is unknown.\n* 203 - COSIGNATORY_NO_PUBLIC_KEY, The cosignatory's public key is unknown.\n* 901 - ACCOUNT_CACHE_ERROR, There was a problem with the account cache.\n* 1000 - PRIVATE_KEY_PUBLIC_KEY_MISMATCH, The public key derived from the private key does not match the given public key.\n* 1001 - PUBLIC_KEY_ADDRESS_MISMATCH, The address derived from the public key does not match the given address.\n* 1002 - NOT_MAIN_NETWORK_ADDRESS, The address does not belong to the main network.\n* 305 - NIS_NOT_AVAILABLE, NCC is disconnected from the network.",
      "enum": [
        106,
        126,
        202,
        203,
        305,
        901,
        1000,
        1001,
        1002
      ]
    },
    "HttpErrorCodes": {
      "type": "integer",
      "format": "int32",
      "description": "Error codes that indicate the reason for failing:  \n* 400 - Bad request, e.g., mandatory parameter is missing or value is not correct.\n* 401 - Unauthorized, Request must be authorized.\n* 404 - Not Found, requested resource not found.",
      "enum": [
        400,
        401,
        404,
        500
      ]
    },
...
    "WalletNamePasswordPair": {
      "properties": {
        "wallet": {
          "type": "string",
          "description": "The name of the wallet."
        },
        "password": {
          "type": "string",
          "description": "The password used for encrypting the wallet."
        }
      },
      "description": "The name and password of a wallet."
    },
...
    "AccountId": {
      "description": "View model that represents an account id information.",
      "properties": {
        "account": {
          "type": "string"
        }
      }
    },
...
    "AccountWalletPasswordRequest": {
      "description": "A request representing an account and a wallet + password.",
      "allOf": [
        {
          "$ref": "#/definitions/AccountId"
        },
        {
          "$ref": "#/definitions/WalletNamePasswordPair"
        }
      ]
    },
...
    "TransferSendRequest": {
      "description": "A request containing all information necessary to create a transfer.",
      "properties": {
        "wallet": {
          "type": "string",
          "description": "The wallet name."
        },
        "type": {
          "type": "string",
          "description": "The type of transfer (multisig or normal)."
        },
        "multisigAccount": {
          "type": "string",
          "description": "The multisig account address."
        },
        "account": {
          "type": "string",
          "description": "The sender account address."
        },
        "recipient": {
          "type": "string",
          "description": "The recipient account address."
        },
        "amount": {
          "type": "integer",
          "format": "int64",
          "description": "The amount."
        },
        "message": {
          "type": "string",
          "description": "The message."
        },
        "encrypt": {
          "type": "integer",
          "format": "int32",
          "description": "Indicates whether the message is encrypted.\n* 0 - not encrypted.\n* 1 - encrypted."
        },
        "hoursDue": {
          "type": "integer",
          "format": "int32",
          "description": "The hours due."
        },
        "password": {
          "type": "string",
          "description": "The password."
        },
        "fee": {
          "type": "integer",
          "format": "int64",
          "description": "The fee."
        },
        "multisigFee": {
          "type": "integer",
          "format": "int64",
          "description": "The multisig transaction fee."
        }
      }
    },
...
    "MultisigSignatureRequest": {
      "description": "A request containing all information necessary to create a transfer.",
      "properties": {
        "wallet": {
          "type": "string",
          "description": "The wallet name."
        },
        "password": {
          "type": "string",
          "description": "The password."
        },
        "account": {
          "type": "string",
          "description": "The sender address."
        },
        "multisigAddress": {
          "type": "string",
          "description": "The multi signature address."
        },
        "innerHash": {
          "type": "string",
          "description": "The inner transaction hash."
        },
        "hoursDue": {
          "type": "integer",
          "format": "int64",
          "description": "The hours due."
        },
        "fee": {
          "type": "integer",
          "format": "int64",
          "description": "The fee."
        }
      }
    },
...
    "PartialTransferInformationRequest": {
      "description": "A request to get information about a partially created transfer.",
      "properties": {
        "recipient": {
          "type": "string",
          "description": "The recipient address."
        },
        "amount": {
          "type": "integer",
          "format": "int64",
          "description": "The amount."
        },
        "message": {
          "type": "string",
          "description": "The message."
        },
        "encrypt": {
          "type": "integer",
          "format": "int32",
          "description": "Indicates whether the message is encrypted.\n* 0 - not encrypted.\n* 1 - encrypted."
        }
      }
    },
    "PartialTransferInformationViewModel": {
      "description": "A request containing all information necessary to create a transfer.",
      "properties": {
        "fee": {
          "type": "integer",
          "format": "int64",
          "description": "The fee."
        },
        "multisigFee": {
          "type": "integer",
          "format": "int64",
          "description": "The fee."
        },
        "encryptionSupported": {
          "type": "integer",
          "format": "int32",
          "description": "Indicates whether the message is encrypted.\n* 0 - not encrypted.\n* 1 - encrypted."
        }
      }
    },
    "PartialSignatureInformationRequest": {
      "description": "A request containing all information necessary to create a transfer.",
      "properties": {
        "multisig": {
          "type": "string",
          "description": "he multisig account address."
        },
        "cosignatory": {
          "type": "string",
          "description": "The cosignatory account address."
        }
      }
    },
...
    "PartialFeeInformationViewModel": {
      "description": "A request containing all information necessary to create a transfer.",
      "properties": {
        "fee": {
          "type": "integer",
          "format": "int64",
          "description": "The fee."
        }
      }
    },
    "TransferImportanceRequest": {
      "description": "A request containing all information necessary to create importance transfer.",
      "allOf": [
        {
          "$ref": "#/definitions/AccountWalletPasswordRequest"
        },
        {
          "properties": {
            "multisigAddress": {
              "type": "string",
              "description": "The multisig account id. Maybe null."
            },
            "type": {
              "type": "integer",
              "format": "int32",
              "description": "The type of transfer (multisig or normal)."
            },
            "hoursDue": {
              "type": "integer",
              "format": "int64",
              "description": "The hours due."
            },
            "fee": {
              "type": "integer",
              "format": "int64",
              "description": "The fee."
            },
            "multisigFee": {
              "type": "integer",
              "format": "int64",
              "description": "the multisig transaction fee."
            },
            "publicKey": {
              "type": "string",
              "description": "The (delegated account) public key. Maybe null."
            }
          }
        }
      ]
    },
...
  },
  "responses": {
    "BadRequest": {
      "description": "Bad request, further information within error response object.",
      "schema": {
        "$ref": "#/definitions/ErrorResponseNcc"
      }
    },
    "Error": {
      "description": "Unexpected error occured.",
      "schema": {
        "$ref": "#/definitions/ErrorResponseHttp"
      }
    }
  }
}